pruebacuatro

+++ apip/Product.cs
@@ -1,5 +1,5 @@
 public class Producto{
     public int Id { get; set; }
     public string? Nombre { get; set; }
-    public decimal Precio { get; set; }
+    public double Precio { get; set; }
 }
